vaccel-client: Fix tensorflow data types
Tensorflow supports only 64bit builds. Use the correct data types and disable build on 32bit architectures

+++ vaccel-client/src/tf_model.rs
@@ -109,7 +109,7 @@ impl VsockClient {
                 let data = e.data;
                 let tensor = ffi::vaccel_tf_tensor_new(
                     dims.len() as i32,
-                    dims.as_ptr() as *mut u32,
+                    dims.as_ptr() as *mut i64,
                     data_type as u32,
                 );
